requirements.txt
langchain
pypdf2
python-dotenv
streamlit
openai==1.14.3
tiktoken
faiss-cpu